==> Analysis/03-compare-patch-protocols.Rmd <==
---
title: "03-compare-patch-protocols"
author: "Ariel Marcy"
date: "2018-10-15"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../aus-rodent-skulls')
```

# Compare big and small patch protocols
Our study asked a methodological question: does it make a difference if we use big areas of patch semi-landmarks versus breaking it up with curve semi-landmarks? Or in other words, will a big patch of landmarks pick up different biological patterns than an area broken up by curve landmarks defined along sutures?

We hypothesized that the big patch protocol would pick up convergent evolution while the small patch + suture curve protocol would better detect phylogentically-conserved shapes. 

Here we test if the shape patterns produced by the two protocols are different in any way.

### Load packages, functions, and data from previous steps
```{r message = FALSE}
# Load libraries (and install if necessary) for landvR package
if(!require(devtools)) install.packages("devtools")
if(!require(dispRity)) install.packages("dispRity")
if(!require(landvR)) install_github("TGuillerme/landvR")
library(data.table)
library(vegan)
library(geomorph)
source("../Functions/utilities.R")  # custom functions
load(file = "../Data/Processed/02-main-data-both-protocols.rda")
```

## Landmark variation heatmaps
Here we use Dr Thomas Guillerme's new package, `landvR` to see how landmarks vary within and between the two protocols. Much of the below was adapted from [this vignette written by Dr Guillerme](https://cdn.rawgit.com/TGuillerme/landvR/8a6a6bd5/inst/vignettes/Landmark_partition_test.html).
```{r}
# Find mean configuration - i.e. mean shape in dataset
big.consensus <- (select.procrustes(big.shape, selector = mean))[[1]]

# Measure spherical coordinates differences from the mean for each specimen for each landmark
diff.from.mean <- coordinates.difference(coordinates = big.shape, reference = big.consensus, type = "spherical")
```

To test the hypothesis that each landmark's variation from the mean configuration is above expected, we can use the PC axes from a PCA.
```{r}
# Ordinate the data
twoD.shape <- two.d.array(big.shape)
ordination <- stats::prcomp(twoD.shape)

# Force the symmetric component of shape into class "gpagen" (required for variation.range)
big.gpagen.shape <- list()
big.gpagen.shape$coords <- big.shape
big.gpagen.shape$consensus <- big.consensus
class(big.gpagen.shape) <- "gpagen"

# Measure extremes of variation from mean on PC1
big.PC1.var <- variation.range(big.gpagen.shape, return.ID = FALSE, axis = 1, ordination = ordination, type = "spherical")
```

The next step is to make a heatmap showing the size of variation and direction of movement for each landmark in the dataset, comparing between the min and max specimens along PC1.
```{r}
# Wrap specimens on the tangent space
big.wrap.PCA <- plotTangentSpace(big.shape, verbose = FALSE)

# Select extreme specimens (the ones that make the warp-meshes) on PC1
big.hypothetical.1 <- big.wrap.PCA$pc.shapes[[1]]  # PC1 min
big.hypothetical.2 <- big.wrap.PCA$pc.shapes[[2]]  # PC1 max

# Plot the range of variation along PC1 using a heat color scheme
PC1.var.plot <- procrustes.var.plot(big.hypothetical.1, big.hypothetical.2, col = heat.colors, col.val = big.PC1.var[, "radius"], labels = F)
```

**Repeat the steps immediately above for PC2**
The pc.shapes[1] and [2] correspond to PC1min and PC1max. Therefore we can adjust the same big.wrap.PCA to plot the heatmaps for PC2. 
```{r}
# Measure extremes of variation from mean on PC2
big.PC2.var <- variation.range(big.gpagen.shape, return.ID = FALSE, axis = 2, ordination = ordination, type = "vector")  # can also try "spherical"

# Select extreme specimens on PC2
big.hypothetical.3 <- big.wrap.PCA$pc.shapes[[3]]  # PC2 min
big.hypothetical.4 <- big.wrap.PCA$pc.shapes[[4]]  # PC2 max

# Plot the range of variation along PC2 using a heat color scheme
PC2.var.plot <- procrustes.var.plot(big.hypothetical.3, big.hypothetical.4, col = heat.colors, col.val = big.PC2.var[, "length"], labels = F)  # use "radius" with "spherical" in line 75
```

## Repeat landmark variation heatmap for small patch protocol PC1 & PC2
```{r}
# Find mean configuration - i.e. mean shape in dataset
sm.consensus <- (select.procrustes(sm.shape, selector = mean))[[1]]

# Measure spherical coordinates differences from the mean for each specimen for each landmark
diff.from.mean <- coordinates.difference(coordinates = sm.shape, reference = sm.consensus, type = "spherical")

# Ordinate the data
twoD.shape <- two.d.array(sm.shape)
ordination <- stats::prcomp(twoD.shape)

# Force the symmetric component of shape into class "gpagen" (required for variation.range)
sm.gpagen.shape <- list()
sm.gpagen.shape$coords <- sm.shape
sm.gpagen.shape$consensus <- sm.consensus
class(sm.gpagen.shape) <- "gpagen"

# Measure extremes of variation from mean on PC1
sm.PC1.var <- variation.range(sm.gpagen.shape, return.ID = FALSE, axis = 1, ordination = ordination, type = "spherical")

# Wrap specimens on the tangent space
sm.wrap.PCA <- plotTangentSpace(sm.shape, verbose = FALSE)

# Select extreme specimens (the ones that make the warp-meshes)
sm.hypothetical.1 <- sm.wrap.PCA$pc.shapes[[1]]
sm.hypothetical.2 <- sm.wrap.PCA$pc.shapes[[2]]

# Plot the range of variation along PC1 using a heat color scheme
PC1.var.plot <- procrustes.var.plot(sm.hypothetical.1, sm.hypothetical.2, col = heat.colors, col.val = sm.PC1.var[, "radius"], labels = F)
```
The heatmaps provide a statistically rigorous way to visualize the relative variation of landmarks across a protocol. In other words, the colors show us which parts of the skull vary most and if the protocols emphasize different patterns. 

**Make heatmap for small patch PC2**
The pc.shapes[1] and [2] correspond to PC1min and PC1max. Therefore we can adjust the same sm.wrap.PCA to plot the heatmaps for PC2. 
```{r}
# Measure extremes of variation from mean on PC2
sm.PC2.var <- variation.range(big.gpagen.shape, return.ID = FALSE, axis = 2, ordination = ordination, type = "vector")  # can also try "spherical"

# Select extreme specimens on PC2
sm.hypothetical.3 <- sm.wrap.PCA$pc.shapes[[3]]  # PC2 min
sm.hypothetical.4 <- sm.wrap.PCA$pc.shapes[[4]]  # PC2 max

# Plot the range of variation along PC2 using a heat color scheme
PC2.var.plot <- procrustes.var.plot(sm.hypothetical.3, sm.hypothetical.4, col = heat.colors, col.val = sm.PC2.var[, "length"], labels = F)  # use "radius" with "spherical" in line 121
```

### Compare histograms of landmark variation by protocol
We can quantify this variation with histograms of the variation for each protocol. 

Big patch protocol:
```{r}
big.var.range <- variation.range(big.gpagen.shape)
hist(big.var.range[, 1])
```

Small patch protocol:
```{r}
sm.var.range <- variation.range(sm.gpagen.shape)
hist(sm.var.range[, 1])
```

## Plot PCAs for each protocol
PCAs allow visual comparison of how differently the patch protocols sort the shape data. To assist, we wrote a function, `PlotByGroup()`, to match colors or other plotting attributes to groups within the dataset. These don't impact the analysis; they are just visualization aids.

Here, we'll use this function to look at the morphospace colored by genus. Since Mastacomys is technically within the genus Pseudomys (Smissen & Rowe 2018), we'll make an "Effective Genus" column.
```{r}
# Make an Effective Genus column to better categorize Mas phylogenetically
big.info$EGenus <- big.info$Genus
big.info$EGenus[which(big.info$Genus == "Mas")] <- "Pse"  # Mas is effectively Pse
sm.info$EGenus <- sm.info$Genus
sm.info$EGenus[which(sm.info$Genus == "Mas")] <- "Pse"  # Mas is effectively Pse

# Colors by EGenus: Con Hyd Leg Lep Mel Mes Mus Not Pog Pse(+Mas) Rat Uro Xer Zyz
col.gen <- c("light green", "red", "cornflowerblue", "dark green", "yellow", "green", "grey", "blue", "magenta", "dark blue", "black", "darkgoldenrod4", "orange", "light blue")

# Colors by specimen, ordered in same way as metadata
sp.col.gen <- PlotByGroup(big.info, "EGenus", col.gen)  # is the same for big and small protocols, tested with all.equal()

# Set color scheme for plot legends, taxa listed in order presented in phylogeny
col.phylo <- c("yellow", "darkgoldenrod4", "dark green", "green", "light green", "cornflowerblue", "light blue", "blue", "dark blue", "red", "orange", "magenta", "grey", "black")  # same as col.gen but in a different order
gen.name.phylo <- c("Mel", "Uro", "Con", "Mes", "Lep", "Leg", "Zyz", "Not", "Pse", "Hyd", "Xer", "Pog", "Mus", "Rat")  # order of genera in phylogeny
```

### PCA plot of big patch protocol
PCAs are very human-friendly way of visualizing shape variation in 2 dimensions. Usually, the proportion of variance for each PC axis is reported in percent.
```{r}
big.pca <- plotTangentSpace(big.shape, groups = sp.col.gen, axis1 = 1, axis2 = 2, verbose = T)

# Flip PC1 and PC2 to show positive shape allometry (larger spp, larger values)
#big.pca$pc.scores[, "PC1"] <- -(big.pca$pc.scores[, "PC1"])
#big.pca$pc.scores[, "PC2"] <- -(big.pca$pc.scores[, "PC2"])

# Get proportion of variance for PC1 and PC2 in format for plots
big.PCs <- big.pca$pc.summary$importance
big.PC1 <- round(big.PCs[2, 1] * 100, digits = 1)  # % with 1 decimal
big.PC2 <- round(big.PCs[2, 2] * 100, digits = 1)

# Pretty plot
plot(x = big.pca$pc.scores[, 1],
     y = big.pca$pc.scores[, 2], 
     xlab = paste("PC1 (", big.PC1, "%)", sep = ""), 
     ylab = paste("PC2 (", big.PC2, "%)", sep = ""),
     xlim = c(-0.11, 0.171),
     asp = TRUE, 
     col = sp.col.gen, 
     bg = sp.col.gen,
     pch = 16, 
     cex = 1.5, 
     cex.axis = 1.3, 
     cex.lab = 1.3)
legend(0.16, 0.055, legend = gen.name.phylo, col = col.phylo, pch = 16, cex = 0.63)
```

### PCA plot of small patch protocol
```{r}
# Plot PCAs
sm.pca <- plotTangentSpace(sm.shape, groups = sp.col.gen, axis1 = 1, axis2 = 2, verbose = T, label = sm.info$CatNum)

# Flip PC1 and PC2 to show positive shape allometry (larger spp, larger values)
#sm.pca$pc.scores[, "PC1"] <- -(sm.pca$pc.scores[, "PC1"])
#sm.pca$pc.scores[, "PC2"] <- -(sm.pca$pc.scores[, "PC2"])

# Get proportion of variance for PC1 and PC2 in format for plots
sm.PCs <- sm.pca$pc.summary$importance
sm.PC1 <- round(sm.PCs[2, 1] * 100, digits = 1)
sm.PC2 <- round(sm.PCs[2, 2] * 100, digits = 1)

# Pretty plot
plot(x = sm.pca$pc.scores[, 1],
     y = sm.pca$pc.scores[, 2], 
     xlab = paste("PC1 (", sm.PC1, "%)", sep = ""), 
     ylab = paste("PC2 (", sm.PC2, "%)", sep = ""),
     xlim = c(-0.11, 0.163),
     asp = T,
     col = sp.col.gen, 
     bg = sp.col.gen,
     pch = 16, 
     cex = 1.5, 
     cex.axis = 1.3, 
     cex.lab = 1.3)
legend(0.16, 0.055, legend = gen.name.phylo, col = col.phylo, pch = 16, cex = 0.63)
```

## Mantel test
A way to test if the two protocols provide a statistically significant different presentations of specimens.
```{r}
mantel(dist(sm.pca$pc.scores), dist(big.pca$pc.scores))
```
The Mantel test finds a highly significant (p < 0.001) relationship between the two PC score distance matrices, with statistic = 0.9989. Meaning they are very similar distance matrices, as expected. 

## Compare performance of patch protocols in ANOVA by genus and centroid size
The patch protocols look very similar on the PCAs, which suggests that they do not capture shape variation differently from one another. To test this, we performed a Procrustes ANOVA on shape ~ centroid size * genus.

### Run shape ~ genus * Csize ANOVAs on each protocol:  
Big patch protocol:
```{r}
big.gdf <- geomorph.data.frame(shape = big.shape, genus = big.info$EGenus, csize = big.info$Csize)
big.procD <- procD.lm(big.shape ~ size * genus, data = gdf)
print(big.procD)

# with procd.allomery
big.Allo <- procD.allometry(shape ~ csize, ~ genus, logsz = TRUE, iter = 999, RRPP = TRUE, print.progress = TRUE, data = big.gdf)
summary(big.Allo)
```

Small patch protocol:
```{r}
gdf <- geomorph.data.frame(genus = sm.info$EGenus, size = sm.info$Csize)
sm.procD <- procD.lm(sm.shape ~ size * genus, data = gdf)
print(sm.procD)

# with procd.allomery
sm.Allo <- procD.allometry(sm.shape ~ csize, ~ genus, logsz = TRUE, iter = 999, RRPP = TRUE, print.progress = TRUE, data = gdf)
summary(sm.Allo)

```
These ANOVAs are also very similar. Therefore, we decided to use only the small patch protocol as this protocol captured more information (sutures versus patch points) and had less asymmetry than the big patch protocol. It also resembles the more common method of landmarking.

### Save intermediate data
Only small patch data from now on.
```{r}
# Remove "sm." from shape and metadata matrix names for increased readability
shape <- sm.shape
info <- sm.info
PCA <- sm.pca
PCs <- sm.PCs

save(shape, info, PCA, PCs, col.gen, sp.col.gen, col.phylo, gen.name.phylo, file = "../Data/Processed/03-sm-PCA-with-color.rda")
```
