==> Analysis/04-compare-patch-protocols.Rmd <==
---
title: "04-calculate-user-error"
author: "Ariel Marcy"
date: "9/26/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../aus-rodent-skulls')
```

### Load packages
```{r message = FALSE}
library(stringr)
library(stringi)
library(data.table)
library(geomorph)
```

### Load data from previous steps
```{r}
load(file = "../Data/Results/01-big-patch-data.rda")
load(file = "../Data/Results/01-sm-patch-data.rda")
load(file = "../Data/Results/02-GPA-data.rda")
load(file = "../Data/Results/03-big-bilat-data.rda")
load(file = "../Data/Results/03-sm-bilat-data.rda")
```

## Separate error measurement datasets from main datasets
Now that the shape data have been processed via GPA and bilateral symmetry, we can separate the data intended to measure user error from the main data for the remaining analyses. 

First, remove the replicates from the main shape data for both small and large patch protocols. Create corresponding metadata tables. These data will be used for the majority of remaining analyses.
```{r}
main <- which(big_bY_info$Rep == "non")
big_shape <- big_bY_ss[, , main]
big_info <- big_bY_info[main, ]

sm_shape <- sm_bY_ss[, , main] # main for small patch protocol is same as above, test with all.equal()
sm_info <- sm_bY_info[main, ]
```

Second, create an error measurement dataset for both protocols that includes replicates and the specimens they duplicated (i.e. each CatNum should have two entries, one labeled "ERR", and one labled "non"). These data will be used to calculate user error. 
```{r}
## Find index for both versions of specimens with replicates using rep_spec list make in script 01.
rep_CatNums <- as.character(str_sub(rep_spec, 9, -1)) # isolate CatNum
error_data <- which(sm_bY_info$CatNum %in% rep_CatNums) # same for small and big patch protocols, can test with all.equal()

## Subset shape and metadata tables for each protcol
big_err_shape <- big_bY_ss[, , error_data]
big_info <- big_bY_info[error_data, ]

sm_err_shape <- sm_bY_ss[, , error_data]
sm_info <- sm_bY_info[error_data, ]
```

## Calculate user error
```{r}

```


### Save intermediate data
```{r}
save(big_shape, big_info, sm_shape, sm_info, file = "../Data/Results/04-data-for-analyses.rda")
```
