==> Analysis/04-plot-exploratory-PCAs.Rmd <==
---
title: "04-plot-exploratory-PCAs"
author: "Ariel Marcy"
date: "2018-10-17"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '../aus-rodent-skulls')
```

# Plotting exploratory PCAs
PCAs provide a human-friendly way to visualize shape data. They are a good way to understand what's going on in the data quickly. However, because they artificially collapse the shape variation into orthogonal PC axes, they must be considered exploratory analyses only.

### Load packages, functions, and data
```{r message = FALSE}
library(stringr)
library(data.table)
library(geomorph)
source("../Functions/utilities.R")  # custom functions
load(file = "../Data/Processed/03-sm-PCA-with-color.rda")
```

### Exploratory analyses with PCAs
Here, we explore the variation in rodent skulls by coloring the PCA plots by genus, taxa, and by traits.

Since we are going to add several columns to the metadata table, we'll remove columns no longer needed past the bilateral symmetry and replicate handling stages. 
```{r}
# Prep metadata for column additions
info <- info[, c(-4, -5, -7)]  # removes now-unnecessary columns
info$Order <- c(1:dim(info)[1])  # preserves specimen order so metadata matches shape data
```

Trait information was gathered from the book, Native Mice and Rats by Dr.s Bill Breed and Fred Ford (2007). 
```{r}
# Add trait data to the metadata table, info and preserve specimen order
traits <- read.csv("../Data/Processed/in_ex_traits.csv", header = TRUE)
traits$Taxa <- paste(str_sub(traits$Genus, 1, 1), str_sub(traits$Species, 1, 3), sep = "_")  # make matching Taxa column
info.traits <- merge(info, traits, by = "Taxa", sort = F)
info.traits <- info.traits[order(info.traits$Order), ] 

# Rename Genus and Species columns so short names have their old column name
names(info.traits)[names(info.traits) == 'Genus.x'] <- 'Genus'
names(info.traits)[names(info.traits) == 'Species.x'] <- 'Species'

# Make effective taxa column to deal with Mastacomys belonging to Pseudomys
info.traits$ETaxa <- info.traits$Taxa
info.traits$ETaxa[which(info.traits$Taxa == "M_fus")] <- "P_fus"
```
Long form Genus and Species columns will keep `Genus.y` and `Species.y`. This will become important for the phylogenetic analyses in script 06.

## Plot PCA by genus for small patch dataset with no landmarking errors
The function `PlotPCA()` quickly plots a PCA with specimens colored by groups and optionally given different point shapes for other attributes. The function also reports PC axis variation in %. 

The chunk below demonstrates how a plot can be exported as a .eps file. These file types can be opened in programs like Adobe Illustrator for minor aesthetic adjustments or be exported to a variety of other filetypes.
```{r}
# Plot PCA by genus and export as .eps into ../Data/Results folder
### setEPS()
### postscript("../Data/Results/PCA_genus.eps")
PlotPCA(PCA, 1, 2, sp.col.gen)
legend(0.144, 0.056, legend = gen.name.phylo, col = col.phylo, pch = 16, cex = 0.78)
### dev.off()
```

### Plot PCA by immigration wave
Here we visualize the same PCA but color it by immigration wave:
1: Australian old endemics (AOEs) by blue
2: Native _Rattus_ by black
3: Invasives (*M musculus, R. rattus, R. norvegicus*) by grey

```{r}
# Create Wave column in traits metadata
info.traits$Wave <- 1  
info.traits$Wave[which(info.traits$Genus == "Rat")] <- 2
info.traits$Wave[which(info.traits$Clade == "invRattus")] <- 3
info.traits$Wave[which(info.traits$Clade == "Mus")] <- 4

# Create color vector
col.wave <- c("blue", "black", "grey", "grey")
sp.col.wave <- PlotByGroup(info.traits, "Wave", col.wave)

# Plot PCA
PlotPCA(PCA, 1, 2, sp.col.wave)
legend(-0.145, -0.055, legend = c("AOEs", "Native Rattus", "Invasives"), col = c("blue", "black", "grey"), pch = 16, cex = 0.78)
text(-0.01, -0.08, "Carnivores", col = "blue")
text(0.105, -0.025, "R. norvegicus", col = "grey")
```

## PCA plot by diet and guild
This PCA plot colors specimen points by diet and gives them a point shape based on guild, i.e. where they spend most of their time: aquatic, arboreal, or terrestrial.

Assign colors by diet and point shapes by guild for PCA.
```{r}
# Colors by Carnivorous, Frugivorous, Granivorous, Herbivorus, Omnivorous
col.feed <- c("red", "darkorange", "gold", "green", "burlywood4")
col.feed <- PlotByGroup(info.traits, "Feeding", col.feed)

# Point key: Aquatic (circle), Arboreal (triangle), Terrestrial (square)
pch.gld <- c(19, 24, 22)
pch.gld <- PlotByGroup(info.traits, "Guild", pch.gld)
```

Plot PC1 vs PC2 and add custom legends.
```{r}
PlotPCA(PCA, 1, 2, col.feed, pch.gld)
legend(0.11, -0.043, legend = c("Aquatic", "Arboreal", "Terrest."), col = "black", border = NULL, pch = c(21, 24, 22))
text(-0.01, -0.08, "Carnivore", col = "red")
text(0.11, -0.02, "Frugivore", col = "darkorange")
text(0.073, 0.045, "Herbivore", col = "green")
text(-0.07, -0.06, "Omnivore", col = "burlywood4")
text(-0.11, 0, "Granivore", col = "gold2")
```

## Morphospace partitioning by habitat
So far, most of the variation in the morphospace could be explained by size and genus affiliation, with different genera of similar sizes overlapping quite a bit. This looks like there isn't much differentiation among Australian rodents, but perhaps the differentiation is hidden by geographical region or partitioning of the environment. 

### Write functions to subset and visualize species morphospace by region
First we will subset the PCA scores by species which inhabit the 7 different ecological regions of Australia. Regions and species distributions therein taken from Breed & Ford 2007. `FoundInRegion()` subsets existing PCA data by species present in given region(s).

Next, we wrote a function, `PointOutDiffSpp()` to give plotting points (pch values) by unique species within each genus.

Finally, we'll combine these functions with a loop to plot PCAs with color by taxa for each region, unique species within each genera having a unique point shape, and a legend for only the species present in each region. 
```{r}
region.names <- c("Savannah", "Arid Zone", "NE Wet Forest", "NE Dry Forest", "Pilbara", "South West", "South East")
pc.scores <- subset(PCA$pc.scores, select = 1:3)  # only 1st 3 PCs
col.egen <- col.gen[-4]  # take out Mas color

# Loop through all 7 regions to plot morphospace of species found in each
for (i in 1:7) {
        # Subset info.traits, PC scores, and color key by region
        in.region <- FoundInRegion(info.traits, regions = i)
        info.region <- info.traits[in.region, ]
        pc.region <- as.data.frame(pc.scores[in.region, ])
        col.taxa.region <- sp.col.gen[in.region]
        
        # Assign points by unique species in a genus
        pch.tax.legend <- PointOutDiffSpp(info.region)
        pch.tax.region <- PlotByGroup(info.region, "ETaxa", pch.tax.legend)
        
        # Plot with fancy (i.e. verbose) legend by:
        info.region$Legend <- paste(str_sub(info.region$EGenus, 1, 3), str_sub(info.region$Species, 1, 3))  # Mus musculus -> "Mus mus"

        # 1) Making list of unique binomial names
        unique.taxa <- sort(unique(info.region$Legend))  # alphabetical order
        legend.genus <- substr(unique.taxa, 1, 3)  # makes string of genera names with repeats when there are unique species in a genus
        
        # 2) Making color key for legend
        genera.in.region <- which(levels(info.traits$EGenus) %in% sort(unique(info.region$EGenus)))
        col.gen.reg <- col.egen[genera.in.region]  # subset col.egen for region
        names(col.gen.reg) <- unique(legend.genus)  # make genus color key with names of unique genera
        col.gen.legend <- col.gen.reg[match(legend.genus, names(col.gen.reg))]
        
        # Write x and y labels with proportion of variance for PCx and PCy
        PCs <- PCA$pc.summary$importance
        PCx.per <- round(PCs[2, 1] * 100, digits = 1)  # % with 1 decimal
        PCx.lab <- paste("PC1 (", PCx.per, "%)", sep = "")
        PCy.per <- round(PCs[2, 2] * 100, digits = 1)
        PCy.lab <- paste("PC2 (", PCy.per, "%)", sep = "")
        
        # Plot PCA with legend and labels
        PCA.plot <- plot(x = pc.region$PC1,
                         y = pc.region$PC2,
                         main = paste(region.names[i], "Region Morphospace" ),
                         xlab = PCx.lab,
                         ylab = PCy.lab,
                         xlim = c(-0.11, 0.18),
                         ylim = c(-0.095, 0.05),
                         col = col.taxa.region, 
                         pch = pch.tax.region, 
                         bg = col.taxa.region,
                         cex = 1.5,
                         cex.axis = 1.3, 
                         cex.lab = 1.3)
        legend(0.147, 0.058, legend = unique.taxa, col = col.gen.legend, border = NULL, pch = pch.tax.legend, pt.bg = col.gen.legend, cex = 0.64)
}
```

### Save intermediate data
```{r}
save(info.traits, col.feed, pch.gld, col.wave, col.egen, sp.col.wave, file = "../Data/Processed/04-sm-data-with-traits.rda")
```
